moving random_split() to random module
